module core base commit
