--- inc/bc_low_power.h
@@ -1,8 +1,0 @@
-#ifndef INC_BC_LOW_POWER_H_
-#define INC_BC_LOW_POWER_H_
-
-void bc_low_power_init_gpio();
-void bc_low_power_init();
-void bc_low_power_enter();
-
-#endif /* INC_BC_LOW_POWER_H_ */
