+++ inc/bc_module_core.h
@@ -1,0 +1,7 @@
+#ifndef INC_BC_MODULE_CORE_H_
+#define INC_BC_MODULE_CORE_H_
+
+void bc_module_core_init();
+void bc_module_core_sleep();
+
+#endif /* INC_BC_MODULE_CORE_H_ */
