+++ stm/src/main.c
@@ -1,24 +1,13 @@
 #include <bc_scheduler.h>
+#include <bc_module_core.h>
 #include <stm32l0xx.h>
 
 void application_init(void);
 bc_tick_t application_task(void *param, bc_tick_t tick_now);
 
-void SystemClock_Config(void);
-void Error_Handler(void);
-
 int main(void)
 {
-    HAL_Init();
-
-    __HAL_RCC_SYSCFG_CLK_ENABLE();
-    __HAL_RCC_PWR_CLK_ENABLE();
-
-    HAL_NVIC_SetPriority(SVC_IRQn, 0, 0);
-    HAL_NVIC_SetPriority(PendSV_IRQn, 0, 0);
-    HAL_NVIC_SetPriority(SysTick_IRQn, 0, 0);
-
-    SystemClock_Config();
+    bc_module_core_init();
 
     bc_scheduler_init();
 
@@ -27,53 +16,6 @@ int main(void)
     bc_scheduler_register(application_task, NULL, 0);
 
     bc_scheduler_run();
-}
-
-void SystemClock_Config(void)
-{
-  RCC_OscInitTypeDef RCC_OscInitStruct;
-  RCC_ClkInitTypeDef RCC_ClkInitStruct;
-  RCC_PeriphCLKInitTypeDef PeriphClkInit;
-
-  __HAL_PWR_VOLTAGESCALING_CONFIG(PWR_REGULATOR_VOLTAGE_SCALE1);
-
-  RCC_OscInitStruct.OscillatorType = RCC_OSCILLATORTYPE_HSI|RCC_OSCILLATORTYPE_HSI48;
-  RCC_OscInitStruct.HSIState = RCC_HSI_ON;
-  RCC_OscInitStruct.HSICalibrationValue = 16;
-  RCC_OscInitStruct.HSI48State = RCC_HSI48_ON;
-  RCC_OscInitStruct.PLL.PLLState = RCC_PLL_ON;
-  RCC_OscInitStruct.PLL.PLLSource = RCC_PLLSOURCE_HSI;
-  RCC_OscInitStruct.PLL.PLLMUL = RCC_PLLMUL_4;
-  RCC_OscInitStruct.PLL.PLLDIV = RCC_PLLDIV_2;
-  if (HAL_RCC_OscConfig(&RCC_OscInitStruct) != HAL_OK)
-  {
-    Error_Handler();
-  }
-
-  RCC_ClkInitStruct.ClockType = RCC_CLOCKTYPE_HCLK|RCC_CLOCKTYPE_SYSCLK
-                              |RCC_CLOCKTYPE_PCLK1|RCC_CLOCKTYPE_PCLK2;
-  RCC_ClkInitStruct.SYSCLKSource = RCC_SYSCLKSOURCE_PLLCLK;
-  RCC_ClkInitStruct.AHBCLKDivider = RCC_SYSCLK_DIV1;
-  RCC_ClkInitStruct.APB1CLKDivider = RCC_HCLK_DIV1;
-  RCC_ClkInitStruct.APB2CLKDivider = RCC_HCLK_DIV1;
-  if (HAL_RCC_ClockConfig(&RCC_ClkInitStruct, FLASH_LATENCY_1) != HAL_OK)
-  {
-    Error_Handler();
-  }
-
-  PeriphClkInit.PeriphClockSelection = RCC_PERIPHCLK_I2C1 | RCC_PERIPHCLK_USB;
-  PeriphClkInit.I2c1ClockSelection = RCC_I2C1CLKSOURCE_PCLK1;
-  PeriphClkInit.UsbClockSelection = RCC_USBCLKSOURCE_HSI48;
-  if (HAL_RCCEx_PeriphCLKConfig(&PeriphClkInit) != HAL_OK)
-  {
-    Error_Handler();
-  }
-
-  HAL_SYSTICK_Config(HAL_RCC_GetHCLKFreq() / 1000);
-
-  HAL_SYSTICK_CLKSourceConfig(SYSTICK_CLKSOURCE_HCLK);
-
-  HAL_NVIC_SetPriority(SysTick_IRQn, 0, 0);
 }
 
 __attribute__((weak)) void application_init(void)
@@ -86,10 +28,4 @@ __attribute__((weak)) bc_tick_t application_task(void *param, bc_tick_t tick_now
     (void) tick_now;
 
     return BC_TICK_INFINITY;
-}
-
-void Error_Handler(void)
-{
-    // TODO Replace
-    for (;;);
 }
